requirements.txt
cat > requirements.txt << 'EOF'
# Core dependencies
nvidia-nemo==2.0.0
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99

# API and UI
fastapi>=0.95.0
uvicorn>=0.22.0
gradio>=3.32.0

# Data processing
datasets>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0

# Deployment
docker>=6.0.0
tensorrt-llm>=0.5.0
EOF